fix: vinxi/react jsx -> js, add types
