+++ packages/vinxi-react/package.json
@@ -10,26 +10,12 @@
   ],
   "types": "./index.d.ts",
   "exports": {
-    ".": "./index.jsx",
-    "./server": "./lib/server.js",
-    "./routes": "./lib/routes.js",
-    "./client": "./lib/app-client-plugin.js",
-    "./lib/*": "./lib/*.js",
-    "./runtime/*": "./runtime/*.js"
+    ".": "./index.jsx"
   },
   "typesVersions": {
     "*": {
       ".": [
         "./index.d.ts"
-      ],
-      "./client": [
-        "./client.d.ts"
-      ],
-      "./server": [
-        "./server.d.ts"
-      ],
-      "./routes": [
-        "./routes.d.ts"
       ]
     }
   },
