+++ packages/vinxi-react/render-asset.js
@@ -1,0 +1,32 @@
+import React from "react";
+
+export function renderAsset({
+	tag,
+	attrs: { key, ...attrs } = {
+		key: undefined,
+	},
+	children,
+}) {
+	switch (tag) {
+		case "script":
+			if (attrs.src) {
+				return createElement("script", { ...attrs, key: attrs.src });
+			} else {
+				return createElement("script", {
+					...attrs,
+					key: key,
+					dangerouslySetInnerHTML: {
+						__html: children,
+					},
+				});
+			}
+		case "link":
+			return createElement("link", { ...attrs, key: key });
+		case "style":
+			return createElement("style", {
+				...attrs,
+				key: key,
+				dangerouslySetInnerHTML: { __html: children },
+			});
+	}
+}
