+++ packages/vinxi-react/tsconfig.json
@@ -6,8 +6,10 @@
 		"allowSyntheticDefaultImports": true,
 		"esModuleInterop": true,
 		"allowJs": true,
-		"noEmit": true,
 		"jsx": "react-jsx",
-		"isolatedModules": true
+		"isolatedModules": true,
+		"skipLibCheck": true,
+		"skipDefaultLibCheck": true,
+		"outDir": "dist"
 	}
 }
